+++ lib/src/constant.dart
@@ -1,0 +1,37 @@
+class KpostalConst {
+  KpostalConst._();
+
+  static const String packageName = 'Kpostal';
+
+  static const String localeKo = 'ko_KR';
+  static const String localeUs = 'en_US';
+
+  static const String postCode = 'zonecode';
+  static const String address = 'address';
+  static const String addressEng = 'addressEnglish';
+  static const String roadAddress = 'roadAddress';
+  static const String roadAddressEng = 'roadAddressEnglish';
+  static const String jibunAddress = 'jibunAddress';
+  static const String jibunAddressEng = 'jibunAddressEnglish';
+  static const String buildingCode = 'buildingCode';
+  static const String buildingName = 'buildingName';
+  static const String apartment = 'apartment';
+  static const String addressType = 'addressType';
+  static const String sido = 'sido';
+  static const String sidoEng = 'sidoEnglish';
+  static const String sigungu = 'sigungu';
+  static const String sigunguEng = 'sigunguEnglish';
+  static const String sigunguCode = 'sigunguCode';
+  static const String roadnameCode = 'roadnameCode';
+  static const String roadname = 'roadname';
+  static const String roadnameEng = 'roadnameEnglish';
+  static const String bcode = 'bcode';
+  static const String bname = 'bname';
+  static const String bname1 = 'bname1';
+  static const String bnameEng = 'bnameEnglish';
+  static const String query = 'query';
+  static const String userSelectedType = 'userSelectedType';
+  static const String userLanguageType = 'userLanguageType';
+  static const String kakaoLatitude = 'kakaoLat';
+  static const String kakaoLongitude = 'kakaoLng';
+}